Replace RouteStop model with CSV fields
This change drops the route_stops table and replaces it with two text
fields:

- bus_stops.bus_route_ids
- bus_routes.bus_stop_ids

These hold CSVs of the IDs of the relevant model lists (i.e. which
routes a bus stop is on, and which stops comprise a given route).

This shitty design is entirely because of Heroku's limit of 10,000
records in a hobby postgres database. The route_stops table had 7,000+
records, so restructuring the data this way means I shouldn't have to
worry about hitting the limit anytime soon.

The code to add route data to the database is woefully inefficient, but
I rarely need to run it, so I don't much care. In future, any data
updates will cause considerable downtime to the app, which could be a
problem if anyone but me is using it. OTOH, running data updates late at
night would be an easy workaround.

fixes #6